+++ Tests/Mockolate.Tests/Web/HttpClientExtensionsTests.Verify.ExpectationMessageTests.cs
@@ -5,79 +5,97 @@ using Mockolate.Web;
 
 namespace Mockolate.Tests.Web;
 
-public sealed class HttpClientExtensionsVerifyExpectationMessageTests
+public sealed partial class HttpClientExtensionsTests
 {
-	[Fact]
-	public async Task DeleteAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
+	public sealed partial class Verify
 	{
-		HttpClient httpClient = HttpClient.CreateMock();
+		public sealed class ExpectationMessageTests
+		{
+			[Fact]
+			public async Task DeleteAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
 
-		void Act()
-			=> httpClient.Mock.Verify.DeleteAsync(It.Matches("*aweXpect*")).AtLeastOnce();
+				void Act()
+				{
+					httpClient.Mock.Verify.DeleteAsync(It.Matches("*aweXpect*")).AtLeastOnce();
+				}
 
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
-	}
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
 
-	[Fact]
-	public async Task DeleteAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
-	{
-		HttpClient httpClient = HttpClient.CreateMock();
+			[Fact]
+			public async Task DeleteAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
 
-		void Act()
-			=> httpClient.Mock.Verify.DeleteAsync(It.IsUri("*aweXpect*")).AtLeastOnce();
+				void Act()
+				{
+					httpClient.Mock.Verify.DeleteAsync(It.IsUri("*aweXpect*")).AtLeastOnce();
+				}
 
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
-	}
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
+
+			[Fact]
+			public async Task PutAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
+
+				void Act()
+				{
+					httpClient.Mock.Verify.PutAsync(It.Matches("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				}
+
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
+
+			[Fact]
+			public async Task PutAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
+
+				void Act()
+				{
+					httpClient.Mock.Verify.PutAsync(It.IsUri("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				}
+
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
 
 #if NET8_0_OR_GREATER
-	[Fact]
-	public async Task PatchAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
-	{
-		HttpClient httpClient = HttpClient.CreateMock();
+			[Fact]
+			public async Task PatchAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
 
-		void Act()
-			=> httpClient.Mock.Verify.PatchAsync(It.Matches("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				void Act()
+				{
+					httpClient.Mock.Verify.PatchAsync(It.Matches("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				}
 
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
-	}
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
 
-	[Fact]
-	public async Task PatchAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
-	{
-		HttpClient httpClient = HttpClient.CreateMock();
+			[Fact]
+			public async Task PatchAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
+			{
+				HttpClient httpClient = HttpClient.CreateMock();
 
-		void Act()
-			=> httpClient.Mock.Verify.PatchAsync(It.IsUri("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				void Act()
+				{
+					httpClient.Mock.Verify.PatchAsync(It.IsUri("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
+				}
 
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
-	}
+				await That(Act).Throws<MockVerificationException>()
+					.WithMessage("*SendAsync*").AsWildcard();
+			}
 #endif
-
-	[Fact]
-	public async Task PutAsync_StringUri_FailingVerification_ShouldMentionSendAsyncInMessage()
-	{
-		HttpClient httpClient = HttpClient.CreateMock();
-
-		void Act()
-			=> httpClient.Mock.Verify.PutAsync(It.Matches("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
-
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
-	}
-
-	[Fact]
-	public async Task PutAsync_Uri_FailingVerification_ShouldMentionSendAsyncInMessage()
-	{
-		HttpClient httpClient = HttpClient.CreateMock();
-
-		void Act()
-			=> httpClient.Mock.Verify.PutAsync(It.IsUri("*aweXpect*"), It.IsAny<HttpContent>()).AtLeastOnce();
-
-		await That(Act).Throws<MockVerificationException>()
-			.WithMessage("*SendAsync*").AsWildcard();
+		}
 	}
 }
